Move Lua native api to dotnethelper
